tune2fs: fix quota enablement regression
Commit dae54ed04c (Avoid dereferencing beyond allocated memory in quota
handling) induced a quota enablement regression in tune2fs.  Attempts
to enable a particular quota type (tune2fs -O project <dev>, for
example) set the appropriate feature flag in the superblock but the
quota file for the type is not initialized.

The regression prevents the current version of xfstests from running
generic/383, 384, 385, and 386 (all project quota tests) with e2fsprogs
1.43.4.  These tests exit with the message "Installed kernel does not
support project quotas", but the actual problem is that the project
quota file is not present.

Signed-off-by: Eric Whitney <enwlinux@gmail.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>