e2fsck: fix reading fscrypt_symlink_data.len
The ciphertext length field stored at the beginning of encrypted symlink
targets is 16-bit.  But e2fsck_pass1_check_symlink() is reading it as
32-bit.  This was apparently left over from an earlier on-disk format
that was not merged.  Fix it.

This bug caused a small proportion of encrypted symlinks with 4092-byte
targets to be considered invalid by e2fsck, but otherwise had no effect.

Fixes: 7bb95ea0c77e ("e2fsck: handle encrypted directories which are indexed using htree")
Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>