e2fsck/recovery.c: sync up with kernel's use of __be32
E2fsprogs as a whole is not sparse-clean, but it does have and
understand the __beXX and __leXX types from the kernel.  The structure
definitions in kernel-jbd.h have been updated to use the __beXX types,
so that recovery.c and revoke.c are more sparse-clean.

This removes a few more unneeded deltas from the kernel's recovery.c.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>