Revert "e2fsck: Change kmem_cache_t to lkmem_cache_t for Solaris"
This is a logical revert of commit 1911bf113ef0, for which the
description reads:

    Solaris polutes the C namespace with kmem_cache_t when
    you include in/netinet.h is included, so rename kmem_cache_t
    to lkmem_cache_t.

Reverting this change allows us to keep e2fsck/revoke.c in sync with
its upstream kernel source of fs/jbd2/revoke.c, and was the last
change required to make the e2fsprogs and kernel versions of revoke.c
to be bit identical.

I've confirmed that this is no longer a problem with OmniOS (an
Illumos / Open Solaris derivative).  It may be a problem with Solaris,
but since I don't have easy access to Solaris, ¯\_(ツ)_/¯

Signed-off-by: Theodore Ts'o <tytso@mit.edu>