e2fsck: don't check for future superblock times if checkinterval == 0
We are no longer enabling periodic file system checks by default in
mke2fs.  The only reason why we force file system checks if the last
mount time or last write time in the superblock is if this might
bypass the periodic file systme checks.  So if the checkinterval is
zero, skip the last mount/write time checks since there's no reason to
force a check just because the system clock is incorrect.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>