e2fsck: restart the full e2fsck run if the bad block inode is invalidate
Previously, we just cleared the bad block list and restarted the inode
scan, but we didn't do a full reset of all of e2fsck's state.  When
code handling this case; we didn't have the framework to do a
restarted run.  Now that we do, we can simply the code and make it
more correct.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>